Stabilize retained output tests
